Align OCI serving mode detection
