+++ wayflowcore/src/wayflowcore/models/ociclientconfig.py
@@ -6,6 +6,7 @@
 
 import logging
 import os
+import re
 import warnings
 from abc import ABC
 from copy import deepcopy
@@ -43,6 +44,24 @@ class ServingMode(str, Enum):
 
     ON_DEMAND = "ON_DEMAND"
     DEDICATED = "DEDICATED"
+
+
+_OCID_RESOURCE_TYPE_RE = re.compile(r"^ocid\d+\.(?P<resource_type>[^.]+)\.")
+_DEDICATED_RESOURCE_TYPES = {"generativeaimodel", "generativeaiendpoint"}
+
+
+def _detect_serving_mode_from_model_id(model_id: str) -> ServingMode:
+    """Infer OCI GenAI serving mode from a model or endpoint identifier.
+
+    Examples
+    --------
+    ``cohere.embed-v4.0`` -> ``ServingMode.ON_DEMAND``
+    ``ocid1.generativeaiendpoint.oc1...`` -> ``ServingMode.DEDICATED``
+    """
+    match = _OCID_RESOURCE_TYPE_RE.match(model_id)
+    if match and match.group("resource_type") in _DEDICATED_RESOURCE_TYPES:
+        return ServingMode.DEDICATED
+    return ServingMode.ON_DEMAND
 
 
 @dataclass
